modified:   firmware/mcu/include/config/GeneratedBuildConfig.h 	modified:   firmware/mcu/include/config/GeneratedLimits.h 	modified:   firmware/mcu/src/core/IntentBuffer.cpp 	modified:   firmware/mcu/src/setup/SetupControlTask.cpp 	modified:   host/mara_host/command/coms/reliable_commander.py 	modified:   host/mara_host/core/_generated_config.py 	modified:   host/mara_host/mcp/_generated_http.py 	modified:   host/mara_host/mcp/_generated_tools.py 	modified:   host/mara_host/services/tooling/backends/_generated_loaders.py

+++ firmware/mcu/src/core/IntentBuffer.cpp
@@ -102,7 +102,7 @@ bool IntentBuffer::consumeDcMotorIntent(uint8_t id, DcMotorIntent& out) {
 // =============================================================================
 
 void IntentBuffer::setStepperIntent(int id, int steps, float speed, uint32_t now_ms) {
-    if (id < 0 || id >= MAX_STEPPER_INTENTS) return;
+    if (id < 0 || id >= static_cast<int>(MAX_STEPPER_INTENTS)) return;
 
     CriticalSection lock(lock_);
     steppers_[id].motor_id = id;
@@ -113,7 +113,7 @@ void IntentBuffer::setStepperIntent(int id, int steps, float speed, uint32_t now
 }
 
 bool IntentBuffer::consumeStepperIntent(int id, StepperIntent& out) {
-    if (id < 0 || id >= MAX_STEPPER_INTENTS) return false;
+    if (id < 0 || id >= static_cast<int>(MAX_STEPPER_INTENTS)) return false;
 
     CriticalSection lock(lock_);
     if (!steppers_[id].pending) {
